Switch to a logical grid
The previous PositionBuffer-based approach tightly coupled display (the
per-frame floating point locations of objects) and collision logic.
That's generally not desirable for puzzle games, and can lead to strange
and subtle bugs when running at low framerates, etc.

Instead, set up a 'logical grid' that we explicitly update during moves.
This has the downside that we have to be careful to remember to invoke it
whenever we're doing strange things with moves, but it makes life much
simpler and easier to reason about in other ways.

I haven't tested this with an extremely large level (ala the Bonfire
Peaks overworld), so I don't know how well it scales. SyncGrid is almost
certainly going to be the problem if so, but it should be fairly
straightforward to make that method smarter about what it changes. (One
could keep a 'dirty set' of objects that have moved/changed & pass that
in to SyncGrid, for example.)